+++ R/pat_loadLatest.R
@@ -7,7 +7,19 @@
 #' @description A pre-generated PurpleAir Timeseries \emph{pat} object will be 
 #' loaded containing data for the most recent 7-day interval.
 #' 
+#' @note Many Purple Air sensor labels have spaces and other special characters
+#' which make for awkward file names. Best practices would suggest creating
+#' file names without special characters by running the \code{label} string
+#' through \code{make.names()} first.
+#'  
+#' This function defaults to generating file names in this manner but allows
+#' users to override this in case some users have a compelling reason to create 
+#' filenames that exactly match the \code{pat$label} of the \emph{pat} object
+#' they contain.
+#' 
 #' @param label PurpleAir sensor 'label'.
+#' @param make.names Logical specifying whether to run 
+#' \code{make.names(label)} when assembilng the file path.
 #' 
 #' @return A PurpleAir Timeseries \emph{pat} object.
 #' 
@@ -23,12 +35,16 @@
 #' }
 
 pat_loadLatest <- function(
-  label = NULL
+  label = NULL,
+  make.names = TRUE
 ) {
   
   # ----- Validate parameters --------------------------------------------------
   
   MazamaCoreUtils::stopIfNull(label)
+  
+  if ( make.names ) 
+    label <- make.names(label)
   
   # ----- Load data from URL or directory --------------------------------------